Adds breakdown to Weekend view
